+++ app/Http/Controllers/CategoryController.php
@@ -43,13 +43,31 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $categoryver = Category::all()->where('categoryName', '==', $request->get('categoryName'));
+        
+        if($categoryver != '[]' ){
+            return "Ya existe una categoría con ese nombre.";
+        }
+
+        if($request->categoryName == NULL or $request->categoryDescription == NULL){
+            return "La categoria y su descripción no pueden ser nulas. Inténtelo denuevo.";
+        }
+        
+        elseif(strlen($request->categoryName) > 20){
+            return "El nombre de la categoría no puede tener mas de 20 caracteres. Inténtelo denuevo.";
+        }
+        
+        elseif(strlen($request->categoryDescription) > 200){
+            return "La descripción de la categoría no puede tener mas de 200 caracteres. Inténtelo denuevo.";
+        }
+
         $category = new Category();
         $category->categoryName = $request->categoryName;
         $category->categoryDescription = $request->categoryDescription;
         $category->visible = $request->visible;
         $category->save();
         return response()->json([
-            "message" => "record created"
+            "message" => "La categoria ha sido creada"
         ], 201);
     }
 
@@ -62,6 +80,10 @@ class CategoryController extends Controller
     public function show($id)
     {
         $category = Category::find($id);
+        
+        if($category == NULL){
+            return "No se ha encontrado la categoría.";
+        }
         return $category;
     }
 
@@ -87,10 +109,28 @@ class CategoryController extends Controller
     {
         $category = Category::findOrFail($id);
 
+        if($category == NULL){
+            return "No se ha encontrado la categoría.";
+        }
+
+        $categoryver = Category::all()->where('categoryName', '==', $request->get('categoryName'));
+        
+        //echo " $categoryver ";
+        if($categoryver != '[]'){
+            //echo " $categoryver ";
+            return "Ya existe una categoría con ese nombre.";
+        }
+
         if ($request->get('categoryName') != NULL){
+            if(strlen($request->categoryName) > 20){
+                return "La categoría no puede tener mas de 20 caracteres. Inténtelo denuevo.";
+            }
             $category->categoryName = $request->get('categoryName');
         }
         if ($request->get('categoryDescription') != NULL){
+            if(strlen($request->categoryDescription) > 200){
+                return "La descripción de la categoría no puede tener mas de 200 caracteres. Inténtelo denuevo.";
+            }
             $category->categoryDescription = $request->get('categoryDescription');
         }
         if ($request->get('visible') != NULL){
@@ -111,6 +151,9 @@ class CategoryController extends Controller
     public function deleteData($id)
     {
         $category = Category::find($id);
+        if($category == NULL){
+            return "No se ha encontrado la categoría.";
+        }
         $category->delete();
         return "La categoría ha sido eliminada";
     }
@@ -119,6 +162,9 @@ class CategoryController extends Controller
     public function deleteVisibility($id)
     {
         $category = Category::find($id);
+        if($category == NULL){
+            return "No se ha encontrado la categoría.";
+        }
         $category->visible = false;
         $category->save();
         return "La categoría ha sido eliminada";
